+++ task/Task23.java
@@ -9,7 +9,7 @@ import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.ArrayWritable;
 import org.apache.hadoop.io.IntWritable;
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
@@ -60,7 +60,7 @@ public class Task23 extends ImprovedTask implements Tool {
 
         Job jobA = Job.getInstance(conf, "Origin-Destination|Carrier Arrival Delay");
         jobA.setOutputKeyClass(Text.class);
-        jobA.setOutputValueClass(DoubleWritable.class);
+        jobA.setOutputValueClass(FloatWritable.class);
 
         jobA.setMapperClass(OriginDestinationCarrierArrDelayMap.class);
         jobA.setReducerClass(ReduceAverage.class);
@@ -90,7 +90,7 @@ public class Task23 extends ImprovedTask implements Tool {
         return jobB.waitForCompletion(true)? 0 : 1;
     }
 
-    public static class OriginDestinationCarrierArrDelayMap extends Mapper<Object, Text, Text, DoubleWritable> {
+    public static class OriginDestinationCarrierArrDelayMap extends Mapper<Object, Text, Text, FloatWritable> {
         @Override
         public void map(Object lineNum, Text value, Context context) throws IOException, InterruptedException {
             String[] row = value.toString().split("\\s");
@@ -98,11 +98,11 @@ public class Task23 extends ImprovedTask implements Tool {
                 String origin = row[5];
                 String destination = row[6];
                 String carrier = row[4];
-                double arrDelay = Double.parseDouble(row[9]);
+                Float arrDelay = Float.parseFloat(row[9]);
                 
                 // paired key = (origin-destination, carrier)
                 String pairedKey = (origin + "_" + destination + "-" + carrier).toUpperCase();
-                context.write(new Text(pairedKey), new DoubleWritable(arrDelay));
+                context.write(new Text(pairedKey), new FloatWritable(arrDelay));
             } catch (Exception e) {
                 // skip on error parsing
             }
